adding logs
