+++ runner/src/server/plugins/engine/models/WebhookModel.ts
@@ -10,7 +10,6 @@ import {
 } from "../../../../../../digital-form-builder/runner/src/server/plugins/engine/components/SelectionControlField";
 import {Field} from "../../../../../../digital-form-builder/runner/src/server/schemas/types";
 import {PageControllerBase} from "../page-controllers/PageControllerBase";
-import {SummaryPageController} from "../page-controllers/SummaryPageController";
 
 export function WebhookModel(model: AdapterFormModel, state: FormSubmissionState) {
     let englishName = `${config.serviceName} ${model.basePath}`;
@@ -23,7 +22,7 @@ export function WebhookModel(model: AdapterFormModel, state: FormSubmissionState
 
     //Getting pages based on the answers in the state & this will allow the user to save all the given answers in the db
     //TODO probable redesign will happen with https://dluhcdigital.atlassian.net/browse/FS-4341 Jira this tries to manage application state based on scenarios
-    const {relevantPages} = getRelevantPagesBasedOnState(state, model);
+    const {relevantPages} = model.getRelevantPagesBasedOnState(state);
 
     questions = relevantPages.map((page) => pagesToQuestions(page, state));
     const fees = FeesModel(model, state);
@@ -35,25 +34,6 @@ export function WebhookModel(model: AdapterFormModel, state: FormSubmissionState
         questions: questions,
         ...(!!fees && {fees}),
     };
-}
-
-function getRelevantPagesBasedOnState(state: FormSubmissionState, model: AdapterFormModel) {
-    let nextPage = model.startPage;
-    const relevantPages: any[] = [];
-    let endPage = null;
-    while (nextPage != null) {
-        if (nextPage.hasFormComponents && nextPage.hasDataInThePage(state)) {
-            relevantPages.push(nextPage);
-        } else if (!nextPage.hasNext && !(nextPage instanceof SummaryPageController)) {
-            endPage = nextPage;
-        }
-        if (nextPage.getNextPage) {
-            nextPage = nextPage.getNextPage(state, true);
-        } else {
-            nextPage = null;
-        }
-    }
-    return {relevantPages, endPage};
 }
 
 function createToFieldsMap(state: FormSubmissionState) {
